refactor: inventory, product 와 연관된 셀러 도메인 병합
