+++ src/main/java/com/onshop/shop/inventory/InventoryRepository.java
@@ -9,8 +9,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import com.onshop.shop.seller.inventory.SellerInventoryResponseDTO;
-
 public interface InventoryRepository extends JpaRepository<Inventory, Long> {
 
 	// 판매자ID 별 재고 조회
