+++ src/main/java/com/onshop/shop/product/ProductsService.java
@@ -20,5 +20,14 @@ public interface ProductsService {
     void getProductDetails(Long productId);
     
     Page<Product> getProductsBySeller(Long sellerId, Pageable pageable);
+    
+    
+    /* 판매자 */
+	List<SellerProductsDTO> getAllProducts(int page, int size); // 모든 상품 조회
+	void registerProducts(List<SellerProductsRequestDTO> products); // 상품 추가(등록)
+    void removeProducts(SellerProductIdsDTO productsIds); 	// 상품 삭제
+    void updateProducts(Long productId, SellerProductsRequestDTO product); // 상품 정보 수정
+    List<SellerProductsDTO> searchProducts(String search, int page, int size);
+    
 
 }
