+++ src/main/java/com/onshop/shop/product/SellerProductIdsDTO.java
@@ -1,4 +1,4 @@
-package com.onshop.shop.seller.products;
+package com.onshop.shop.product;
 
 import java.util.List;
